Fix issues with timeouts on all requests when using curl 8.20

+++ src/sevents.h
@@ -127,7 +127,7 @@ static void send_now_playing(evutil_socket_t fd, short event, void *data)
     // TODO(marius): this requires the number of tracks to be passed down, to avoid dependency on arrlen
     api_request_do(scrobbler, tracks, 1, now_playing_is_valid, api_build_request_now_playing);
 
-    if (track->position + NOW_PLAYING_DELAY < (double)track->length) {
+    if (track->position + NOW_PLAYING_DELAY < track->length) {
         add_event_now_playing(player, track, NOW_PLAYING_DELAY);
     }
 }
